@@ -4,40 +4,26 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=7.2.5",
+        "php": ">=8.2",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "composer/package-versions-deprecated": "1.11.99.2",
-        "doctrine/annotations": "^1.0",
-        "doctrine/doctrine-bundle": "^2.4",
-        "doctrine/doctrine-migrations-bundle": "^3.1",
-        "doctrine/orm": "^2.9",
-        "nelmio/api-doc-bundle": "^4.4",
-        "nelmio/cors-bundle": "^2.1",
-        "phpdocumentor/reflection-docblock": "^5.2",
-        "symfony/asset": "5.3.*",
-        "symfony/console": "5.3.*",
-        "symfony/dotenv": "5.3.*",
-        "symfony/flex": "^1.3.1",
-        "symfony/framework-bundle": "5.3.*",
-        "symfony/property-access": "5.3.*",
-        "symfony/property-info": "5.3.*",
-        "symfony/proxy-manager-bridge": "5.3.*",
-        "symfony/runtime": "5.3.*",
-        "symfony/security-bundle": "5.3.*",
-        "symfony/serializer": "5.3.*",
-        "symfony/twig-bundle": "5.3.*",
-        "symfony/yaml": "5.3.*",
-        "web-token/jwt-checker": "^2.2",
-        "web-token/jwt-core": "^2.2",
-        "web-token/jwt-key-mgmt": "^2.2",
-        "web-token/jwt-signature": "^2.2",
-        "web-token/jwt-signature-algorithm-rsa": "^2.2"
+        "doctrine/doctrine-bundle": "^2.11",
+        "doctrine/doctrine-migrations-bundle": "^3.3",
+        "doctrine/orm": "^2.17",
+        "lexik/jwt-authentication-bundle": "^2.20",
+        "symfony/console": "7.0.*",
+        "symfony/dotenv": "7.0.*",
+        "symfony/flex": "^2",
+        "symfony/framework-bundle": "7.0.*",
+        "symfony/runtime": "7.0.*",
+        "symfony/security-bundle": "7.0.*",
+        "symfony/yaml": "7.0.*"
     },
     "config": {
-        "optimize-autoloader": true,
-        "preferred-install": {
-            "*": "dist"
+        "allow-plugins": {
+            "php-http/discovery": true,
+            "symfony/flex": true,
+            "symfony/runtime": true
         },
         "sort-packages": true
     },
@@ -54,7 +40,12 @@
     "replace": {
         "symfony/polyfill-ctype": "*",
         "symfony/polyfill-iconv": "*",
-        "symfony/polyfill-php72": "*"
+        "symfony/polyfill-php72": "*",
+        "symfony/polyfill-php73": "*",
+        "symfony/polyfill-php74": "*",
+        "symfony/polyfill-php80": "*",
+        "symfony/polyfill-php81": "*",
+        "symfony/polyfill-php82": "*"
     },
     "scripts": {
         "auto-scripts": {
@@ -67,8 +58,7 @@
         "post-update-cmd": [
             "@auto-scripts"
         ],
-        "unit:test": "php bin/phpunit tests --colors=always --testdox",
-        "analyse": ".\\vendor\\bin\\phpstan analyse src tests --ansi || ./vendor/bin/phpstan analyse src tests --ansi"
+        "phpstan": "vendor/bin/phpstan analyse --ansi"
     },
     "conflict": {
         "symfony/symfony": "*"
@@ -76,15 +66,11 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.3.*"
+            "require": "7.0.*"
         }
     },
     "require-dev": {
-        "phpstan/phpstan": "^0.12.94",
-        "phpunit/phpunit": "^9.5",
-        "symfony/browser-kit": "^5.3",
-        "symfony/css-selector": "^5.3",
-        "symfony/maker-bundle": "^1.33",
-        "symfony/phpunit-bridge": "^5.3"
+        "phpstan/phpstan": "^1.10",
+        "symfony/maker-bundle": "^1.53"
     }
 }
