@@ -1,37 +1,25 @@
 <?php
 
-// src/EventListener/ExceptionListener.php
 namespace App\EventListener;
 
+use App\Factory\SwoopyResponseFactory;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
 use Symfony\Component\HttpKernel\Exception\HttpExceptionInterface;
 
 class ExceptionListener
 {
-    public function onKernelException(ExceptionEvent $event)
+    public function onKernelException(ExceptionEvent $event): void
     {
         $exception = $event->getThrowable();
-        $response = new Response();
+        $headers = [];
+        $statusCode = Response::HTTP_INTERNAL_SERVER_ERROR;
 
         if ($exception instanceof HttpExceptionInterface) {
-            $response->setStatusCode($exception->getStatusCode());
-            $response->headers->replace($exception->getHeaders());
-        } else {
-            $response->setStatusCode(Response::HTTP_INTERNAL_SERVER_ERROR);
+            $statusCode = $exception->getStatusCode();
+            $headers = $exception->getHeaders();
         }
 
-        $message = [
-            'status' => 'ERROR',
-            'message' => $exception->getMessage(),
-            'code' => $response->getStatusCode(),
-        ];
-
-        if ($_ENV['APP_ENV'] === 'dev') {
-            $message['stack'] = $exception->getTrace();
-        }
-
-        $response->setContent(json_encode($message));
-        $event->setResponse($response);
+        $event->setResponse(SwoopyResponseFactory::createFailureResponse($exception, $statusCode, $headers));
     }
 }
